# dict: Use DK_LOG_SIZE in hot loop.