DK_LOG_SIZE(key) < 8 is faster than DK_SIZE(key) <= 0xff, at least on
GCC.