@markshannon 2022-02-18
Apart from one variable name, looks good.